/Dev 4.x nft

+++ src/ethereum/ethereum.h
@@ -10,5 +10,5 @@
 #include <ethereum/rlp.h>
 #include <amount.h>
 bool VerifyProof(dev::bytesConstRef path, const dev::RLP& value, const dev::RLP& parentNodes, const dev::RLP& root); 
-bool parseEthMethodInputData(const std::vector<unsigned char>& vchInputExpectedMethodHash,  const uint8_t& nERC20Precision, const uint8_t& nLocalPrecision, const std::vector<unsigned char>& vchInputData, CAmount& outputAmount, uint32_t& nAsset, std::string& witnessAddress);
+bool parseEthMethodInputData(const std::vector<unsigned char>& vchInputExpectedMethodHash,  const uint8_t& nERC20Precision, const uint8_t& nLocalPrecision, const std::vector<unsigned char>& vchInputData, CAmount& outputAmount, uint64_t& nAsset, std::string& witnessAddress);
 #endif // SYSCOIN_ETHEREUM_ETHEREUM_H
